Refactor AccountProvider to hook
